Update `display-name` and remove `description` from `.astro-registry.yaml`
The current value of the `display-name` is not the best value to use for a given provider. Ideally a display name on the Registry for a provider should not include "Airflow" or "Provider" but simply the name of the service/integration for which the provider is associated.  Also, the description is not used for providers and can be removed.